efesto/efesto-core/efesto-common-api/src/main/java/org/kie/efesto/common/api/utils/FileNameUtils.java
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */
package org.kie.efesto.common.api.utils;

import java.io.File;

public class FileNameUtils {
    private FileNameUtils() {
    }

    public static String getFileName(String source) {
        return source.contains(File.separator) ?
                source.substring(source.lastIndexOf(File.separatorChar) + 1) : source;
    }

    public static String getSuffix(String fileName) {
        return fileName.substring(fileName.lastIndexOf('.') + 1);
    }

    public static String removeSuffix(String fileName) {
        return fileName.substring(0, fileName.lastIndexOf('.'));
    }
}
